Use proper longValue() when accessing workflow time for comparisons

+++ wb-wf/src/main/java/org/ihtsdo/ace/task/search/ModelerLastWorkflowHistory.java
@@ -98,7 +98,7 @@ public class ModelerLastWorkflowHistory extends AbstractWorkflowHistorySearchTes
 		WorkflowHistoryJavaBean current = null;
 
 		for (WorkflowHistoryJavaBean wfHistoryItem : wfHistory) {
-			if (current == null || wfHistoryItem.getEffectiveTime() > current.getEffectiveTime())
+			if (current == null || wfHistoryItem.getEffectiveTime().longValue() > current.getEffectiveTime().longValue())
 				current = wfHistoryItem;
 		}
